requirements.txt
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.18.0
timm>=0.5.4
datasets>=2.0.0
pillow>=9.0.0
huggingface-hub>=0.4.0
numpy>=1.20.0
tqdm>=4.62.0
matplotlib>=3.5.0
tensorboard>=2.8.0
pycocotools>=2.0.4
einops>=0.4.1
scipy>=1.7.3
